extract serializers to seperate package (#457)
* extract serializers to seperate package

* clean pubspec

* compat with new dio version, return uint8list in test mocks instead of list<int>
