+++ codegen/end_to_end_test_tristate/pubspec.yaml
@@ -9,10 +9,11 @@ dependencies:
   built_collection: ^5.0.0
   built_value: ^8.0.6
   gql_exec: ^1.0.0
-  gql_build: ^0.8.0
-  gql_code_builder: ^0.7.1
+  gql_code_builder_serializers: ^0.1.0
 dev_dependencies: 
   build: ^2.0.0
   build_runner: ^2.0.0
   test: ^1.16.8
   gql_tristate_value: ^1.0.0
+  gql_build: ^0.8.0
+  gql_code_builder: ^0.7.1
