+++ codegen/gql_build/pubspec.yaml
@@ -14,7 +14,8 @@ dependencies:
   dart_style: ^2.2.4
   glob: ^2.0.0
   gql: ^1.0.0
-  gql_code_builder: ^0.10.0
+  gql_code_builder: ^0.11.0
+  gql_code_builder_serializers: ^0.1.0
   path: ^1.8.0
   yaml: ^3.1.0
   gql_tristate_value: ^1.0.0
