+++ codegen/gql_code_builder/pubspec.yaml
@@ -1,5 +1,5 @@
 name: gql_code_builder
-version: 0.11.0
+version: 0.12.0
 description: Dart code builders taking *.graphql documents and SDL to build useful classes.
 repository: https://github.com/gql-dart/gql
 environment: 
@@ -14,6 +14,7 @@ dependencies:
   gql_exec: ^1.0.0
   path: ^1.8.0
   gql_tristate_value: ^1.0.0
+  gql_code_builder_serializers: ^0.1.0
 dev_dependencies: 
   build_runner: ^2.1.0
   gql_pedantic: ^1.0.2
