+++ codegen/gql_code_builder_serializers/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 0.1.0
+
+- Initial version
