+++ codegen/gql_code_builder_serializers/README.md
@@ -1,0 +1,6 @@
+# gql_code_builder_serializers
+
+Common Serializers for `gql_code_builder`.
+This package should be added as a dependency for packages
+that generate code using `gql_code_builder`.
+
