+++ codegen/gql_code_builder_serializers/lib/gql_code_builder_serializers.dart
@@ -1,0 +1,4 @@
+export 'src/serializers/default_scalar_serializer.dart';
+export 'src/serializers/json_serializer.dart';
+export 'src/serializers/operation_serializer.dart';
+export 'src/serializers/inline_fragment_serializer.dart';
