+++ codegen/gql_code_builder_serializers/pubspec.yaml
@@ -1,0 +1,17 @@
+name: gql_code_builder_serializers
+version: 0.1.0
+description: Common serializers for gql_code_builder
+repository: https://github.com/gql-dart/gql
+environment: 
+  sdk: ^3.0.0
+dependencies: 
+  gql: ^1.0.0
+  built_value: ^8.0.0
+  gql_exec: ^1.0.0
+dev_dependencies: 
+  lints: ^3.0.0
+  test: ^1.24.0
+topics: 
+  - graphql
+  - gql
+  - codegen
